First go at dropping macros implementing function evaluate in BasicGrammar_inl.h
So far :
  - create a commonGramar base class for the Grammar class in order to be able to specialize easily the templated Grammar class
  - specialized the Grammar class for std::vector
  - dropped the macros concerning vector grammars, as they are replaced by the specialization

Note that the specialization is much shorter in code and much more generic than the macros, as the compiler will generate appropriate Grammars for new type without any change of code
